feat(portal): add deep-linkable /portal/requests route
The request-access form was a portal-internal modal with no URL, so the
Solaris companion app could only link to the portal root and the visitor
had to click "Don't have an account yet?" first.

Extract the modal body into RequestAccessDialog so it can be mounted
already open, and add /portal/requests, which renders it pre-opened for
anonymous visitors. The new route reuses the same helpers as /portal:
isPortalBlockedForRequest for the LAN-only gate (same notice, now shared
in PortalLanOnlyNotice), verifyAutheliaSession — a signed-in visitor is
redirected to /portal rather than shown a form for an account they
already have — and AccessRequestStatusCTA, so a visitor with a pending
request sees its status instead of a second blank form.

No proxy change needed: pages are not session-gated, and the apex/www
rewrite passes /portal* paths through untouched.

Closes #2405
